archive_unused/enhanced_requirements.txt
# Enhanced Production Requirements
# Core Dependencies
weave==0.50.7
wandb>=0.16.0
openai>=1.0.0
streamlit>=1.28.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web and APIs
requests
aiohttp>=3.8.0
flask>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
plotly>=5.0.0

# LangChain Framework
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.30

# MCP (Model Context Protocol)
mcp>=0.1.0
httpx>=0.25.0
websockets>=11.0

# Enhanced Agent Capabilities
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Optional Extensions
# openpipe-ai